fixed Docker bug and made some general good practice improvements
