+++ .gitignore
@@ -1,3 +1,6 @@
 .env
 /build/
 /dist/
+
+.vscode/
+node_modules/
