+++ prompt.ts
@@ -4,7 +4,7 @@ Help me to create a repository containing the code for:
 
 The project name is: {REPOSITORY_NAME}
 
-Give me instructions to generate complete, working code for this project, but use only shell commands. No need to give further explanation. All of your output should be given as a single /bin/sh script.
+Give me instructions to generate complete, working code for this project, but use only shell commands. No need to give further explanation. All of your output must be given as a single /bin/sh script.
 
 For example, when want to edit a file, use the command \`echo "file contents" > filename.ext\`. Create intermediate directories before writing files.
 
@@ -12,7 +12,7 @@ At the beginning of the script, create a git repo in the current directory befor
 
 Assume all commands will be run on a clean \`{BASE_IMAGE}\` system with no packages installed.
 
-Follow each command in the instructions with a git commit command including a detailed commit message. Follow each file change in the instructions with a git commit command including a detailed commit message.
+Follow each command in the instructions with a git commit command including a detailed commit message.
 
 The repository should also contain a helpful README.md file. The README should include a high-level descrition of the code, a list of software needed to run the code, and basic instructions on how to run the application. The README should not include any text or information about a software license. This project has no license.
